+++ Source/z88dk_target/src/sys/Types.c
@@ -3,11 +3,11 @@
 const char8_t NewLineChar = '\n';
 const char8_t CarriageReturnChar = '\r';
 
-const uint8_t SmallArray8_size = sizeof(SmallArray8);
+const uint8_t SmallArray8_size = (uint8_t)sizeof(SmallArray8);
 const uint8_t SmallArray8_MaxCapacity = uint8_max;
 
-const uint8_t Array8_size = sizeof(Array8);
+const uint8_t Array8_size = (uint8_t)sizeof(Array8);
 const uint16_t Array8_MaxCapacity = uint16_max;
 
-const uint8_t Collection8_size = sizeof(Collection8);
+const uint8_t Collection8_size = (uint8_t)sizeof(Collection8);
 const uint16_t Collection8_MaxCapacity = uint16_max;
